Move aggregate type resolution to util class
Move aggregate type resolution to util class. This is a bit cleaner than sharing the concern within the AnnotatedAggregateMetaModelFactory

#4644
